test: simplify archive and auth proofs
