+++ packages/backend/convex/contentRelease/archive/retention.test.ts
@@ -1,8 +1,6 @@
 // @vitest-environment node
 
-import { createHash } from "node:crypto";
 import { afterEach, beforeEach, describe, expect, it } from "@effect/vitest";
-import { CONTENT_RUNTIME_ARCHIVE_CONTENT_TYPE } from "@repo/backend/content/archive";
 import {
   CONTENT_RUNTIME_ARCHIVE_CLAIM_PATH,
   CONTENT_RUNTIME_ARCHIVE_FINALIZE_PATH,
@@ -12,92 +10,56 @@ import type { Id } from "@repo/backend/convex/_generated/dataModel";
 import { MAX_CONTENT_RUNTIME_ARCHIVES } from "@repo/backend/convex/contentRelease/archive/spec";
 import { ROLLBACK_RETENTION_MS } from "@repo/backend/convex/contentRelease/spec";
 import { createConvexTestWithBetterAuth } from "@repo/backend/convex/test.helpers";
-import { storeArchiveFixture } from "@repo/backend/test/archive";
-
-const ARCHIVE_TOKEN = "technical-archive-token";
-type RuntimeTest = ReturnType<typeof createConvexTestWithBetterAuth>;
-
-function identity(index: number) {
-  return {
-    runtimeSelectionHash: index.toString(16).padStart(64, "0"),
-    runtimeSchemaFingerprint: "e".repeat(64),
-  };
-}
-
-function sourceStateHash(index: number) {
-  return (index + 1000).toString(16).padStart(64, "0");
-}
-
-function claimId(index: number) {
-  return `00000000-0000-4000-8000-${String(index).padStart(12, "0")}`;
-}
-
-function write(target: RuntimeTest, path: string, body: unknown) {
-  return target.fetch(path, {
-    body: JSON.stringify(body),
-    headers: {
-      "content-type": "application/json",
-      "x-nakafa-archive-token": ARCHIVE_TOKEN,
-    },
-    method: "POST",
-  });
-}
+import {
+  type ArchiveTest,
+  claimId,
+  clearEnvironment,
+  hash,
+  identity,
+  setEnvironment,
+  sourceHash,
+  storeArchiveFixture as store,
+  write,
+} from "@repo/backend/test/archive";
 
 async function insertArchive(
-  target: RuntimeTest,
+  target: ArchiveTest,
   index: number,
   createdAt: number
 ) {
   const value = `retained-runtime-archive-${index}`;
-  const archiveSha256 = createHash("sha256").update(value).digest("hex");
-  const storageId = await storeArchiveFixture(
-    target,
-    value,
-    CONTENT_RUNTIME_ARCHIVE_CONTENT_TYPE
-  );
+  const archiveSha256 = hash(value);
+  const storageId = await store(target, value);
   await target.run((ctx) =>
     ctx.db.insert("contentRuntimeArchives", {
       ...identity(index),
       archiveSha256,
       byteLength: Buffer.byteLength(value),
       createdAt,
-      sourceStateHash: sourceStateHash(index),
+      sourceStateHash: sourceHash(index),
       storageId,
     })
   );
   return storageId;
 }
 
-async function storeNewArchive(target: RuntimeTest, index: number) {
+async function storeNewArchive(target: ArchiveTest, index: number) {
   const value = `new-runtime-archive-${index}`;
   const input = { ...identity(index), claimId: claimId(index) };
   await write(target, CONTENT_RUNTIME_ARCHIVE_CLAIM_PATH, input);
-  const storageId = await storeArchiveFixture(
-    target,
-    value,
-    CONTENT_RUNTIME_ARCHIVE_CONTENT_TYPE
-  );
+  const storageId = await store(target, value);
   const response = await write(target, CONTENT_RUNTIME_ARCHIVE_FINALIZE_PATH, {
     ...input,
-    archiveSha256: createHash("sha256").update(value).digest("hex"),
+    archiveSha256: hash(value),
     byteLength: Buffer.byteLength(value),
-    sourceStateHash: sourceStateHash(index),
+    sourceStateHash: sourceHash(index),
     storageId,
   });
   expect(response.status).toBe(200);
 }
 
-beforeEach(() => {
-  process.env.CONTENT_ARCHIVE_TOKEN = ARCHIVE_TOKEN;
-  process.env.CONTENT_RUNTIME_TOKEN = "technical-runtime-token";
-  process.env.POLAR_WEBHOOK_SECRET = "technical-webhook-secret";
-});
-
-afterEach(() => {
-  delete process.env.CONTENT_ARCHIVE_TOKEN;
-  delete process.env.CONTENT_RUNTIME_TOKEN;
-  delete process.env.POLAR_WEBHOOK_SECRET;
-});
+beforeEach(setEnvironment);
+afterEach(clearEnvironment);
 
 describe("content runtime archive retention", () => {
   it("preserves the current archive when its identity remains unchanged", async () => {
